test: gso 캡션 attr bit29 계약을 통합 테스트로 옮기고 중복 정의 제거
unit tier 게이트는 source 측 단위 테스트 증가를 금지한다
(src/serializer/control.rs::tests 33 > 31). 새 gso 계약을 devel 에 이미 있는
tests/cases/issue_5136_caption_attr_bit29.rs 로 옮기고, 표 계약은 그 파일에
중복 추가하지 않는다. ctrl_attr 헬퍼로 tbl/gso 를 공용화한다.

devel 병합으로 apply_caption_attr_bit 가 두 번 정의돼 컴파일이 깨지던 것도
정리한다 — 남은 정의는 pack_common_attr_bits 의 doc 블록 밖에 둔다.
