+++ src/serializer/control/tests.rs
@@ -371,98 +371,6 @@ fn test_roundtrip_table() {
     }
 }
 
-/// [일곱 번째 계약] 표의 캡션 유무가 개체 공통속성 attr **bit 29**(캡션 존재
-/// 플래그)에 반영돼야 한다. 어긋나면 한글 2022 가 CTRL_HEADER 직후의 캡션
-/// LIST_HEADER 유무를 오판해 문서 전체 개방을 거부한다(크롤 빈티지 40429 표
-/// COM 이등분 실측, 양방향 반증). 종전 HWP3/HWPX 출처 표는 캡션을 방출하면서
-/// 이 비트를 안 켜 개방 거부됐다.
-#[test]
-fn captioned_table_sets_common_attr_bit29() {
-    const CAPTION_BIT: u32 = 1 << 29;
-    let cell = Cell {
-        col: 0,
-        row: 0,
-        col_span: 1,
-        row_span: 1,
-        width: 10000,
-        height: 5000,
-        border_fill_id: 1,
-        paragraphs: vec![Paragraph {
-            char_count: 1,
-            char_offsets: vec![],
-            char_shapes: vec![CharShapeRef {
-                start_pos: 0,
-                char_shape_id: 0,
-            }],
-            line_segs: vec![LineSeg {
-                text_start: 0,
-                ..Default::default()
-            }],
-            ..Default::default()
-        }],
-        ..Default::default()
-    };
-    let make_table = |caption: Option<Caption>| Table {
-        row_count: 1,
-        col_count: 1,
-        row_sizes: vec![1],
-        border_fill_id: 1,
-        cells: vec![cell.clone()],
-        caption,
-        ..Default::default()
-    };
-    // 직렬화된 섹션 레코드에서 'tbl ' CTRL_HEADER 의 attr(payload 0..4)을 읽는다.
-    fn tbl_ctrl_attr(bytes: &[u8]) -> u32 {
-        let mut off = 0usize;
-        while off + 4 <= bytes.len() {
-            let h = u32::from_le_bytes(bytes[off..off + 4].try_into().unwrap());
-            let tag = h & 0x3ff;
-            let mut sz = ((h >> 20) & 0xfff) as usize;
-            let mut body = off + 4;
-            if sz == 0xfff {
-                sz = u32::from_le_bytes(bytes[body..body + 4].try_into().unwrap()) as usize;
-                body += 4;
-            }
-            // CTRL_HEADER(tag 71) 의 ctrl_id 는 payload 0..4 (LE 로 'tbl ' = 0x74626c20)
-            if tag == 71 && bytes[body..body + 4] == *b" lbt" {
-                return u32::from_le_bytes(bytes[body + 4..body + 8].try_into().unwrap());
-            }
-            off = body + sz;
-        }
-        panic!("'tbl ' CTRL_HEADER 를 찾지 못함");
-    }
-    let with = serialize_section(&Section {
-        paragraphs: vec![Paragraph {
-            char_count: 2,
-            controls: vec![Control::Table(Box::new(make_table(Some(caption_fixture(
-                "표 캡션",
-            )))))],
-            ..Default::default()
-        }],
-        raw_stream: None,
-        ..Default::default()
-    });
-    let without = serialize_section(&Section {
-        paragraphs: vec![Paragraph {
-            char_count: 2,
-            controls: vec![Control::Table(Box::new(make_table(None)))],
-            ..Default::default()
-        }],
-        raw_stream: None,
-        ..Default::default()
-    });
-    assert_ne!(
-        tbl_ctrl_attr(&with) & CAPTION_BIT,
-        0,
-        "캡션 있는 표는 attr bit29 가 켜져야 한다"
-    );
-    assert_eq!(
-        tbl_ctrl_attr(&without) & CAPTION_BIT,
-        0,
-        "캡션 없는 표는 attr bit29 가 꺼져야 한다"
-    );
-}
-
 /// AutoNumber 라운드트립
 #[test]
 fn test_roundtrip_auto_number() {
@@ -1114,63 +1022,6 @@ fn issue2715_rectangle_caption_roundtrips() {
     assert_eq!(cap.spacing, 850, "캡션-틀 간격 보존");
     assert_eq!(cap.max_width, 30000, "캡션 최대 폭 보존");
     assert!(cap.include_margin, "include_margin 보존");
-}
-
-/// [일곱 번째 계약의 gso 확장] 캡션 달린 gso 도형(사각형·OLE 등)도 개체
-/// 공통속성 attr **bit 29**(캡션 존재 플래그)를 켜야 한다. 종전에는 표·그림만
-/// 반영하고 gso 도형 arm 은 빠져, 캡션 달린 OLE 가 든 변환본을 한글 2022 가
-/// 거부했다(크롤 빈티지 15456 OLE 객체 COM 이등분: gso CTRL_HEADER attr bit29
-/// 로 확정 — z-order·instance_id 는 무관, 단독 반증).
-#[test]
-fn captioned_gso_shape_sets_common_attr_bit29() {
-    const CAPTION_BIT: u32 = 1 << 29;
-    // 직렬화된 섹션에서 첫 CTRL_GEN_SHAPE('gso ') 의 attr 을 읽는다.
-    fn gso_ctrl_attr(bytes: &[u8]) -> u32 {
-        let mut off = 0usize;
-        while off + 4 <= bytes.len() {
-            let h = u32::from_le_bytes(bytes[off..off + 4].try_into().unwrap());
-            let tag = h & 0x3ff;
-            let mut sz = ((h >> 20) & 0xfff) as usize;
-            let mut body = off + 4;
-            if sz == 0xfff {
-                sz = u32::from_le_bytes(bytes[body..body + 4].try_into().unwrap()) as usize;
-                body += 4;
-            }
-            if tag == 71 && bytes[body..body + 4] == *b" osg" {
-                return u32::from_le_bytes(bytes[body + 4..body + 8].try_into().unwrap());
-            }
-            off = body + sz;
-        }
-        panic!("'gso ' CTRL_HEADER 를 찾지 못함");
-    }
-    let make = |caption: Option<Caption>| {
-        let rect = RectangleShape {
-            drawing: DrawingObjAttr {
-                caption,
-                ..Default::default()
-            },
-            ..Default::default()
-        };
-        serialize_section(&Section {
-            paragraphs: vec![Paragraph {
-                char_count: 2,
-                controls: vec![Control::Shape(Box::new(ShapeObject::Rectangle(rect)))],
-                ..Default::default()
-            }],
-            raw_stream: None,
-            ..Default::default()
-        })
-    };
-    assert_ne!(
-        gso_ctrl_attr(&make(Some(caption_fixture("도형 캡션")))) & CAPTION_BIT,
-        0,
-        "캡션 있는 gso 도형은 attr bit29 가 켜져야 한다"
-    );
-    assert_eq!(
-        gso_ctrl_attr(&make(None)) & CAPTION_BIT,
-        0,
-        "캡션 없는 gso 도형은 attr bit29 가 꺼져야 한다"
-    );
 }
 
 /// [#2715] 묶음(`$con`) 캡션 왕복 — 한컴 `samples/draw-group.hwp` 실측 대응.
